feat(admin): rating soft-delete/restore uçları - RatingRepository: providerId lookup, softDelete/restore JPQL - RatingAdminService: transactional + cache evict - RatingAdminController: DELETE /api/admin/ratings/{id}, POST /{id}/restore - Avg/count triggerlarıyla uyumlu; cache invalidation eklendi

+++ src/main/kotlin/com/furkan/sanayi/dto/RatingDto.kt
@@ -4,7 +4,7 @@ import java.time.Instant
 
 data class RatingDto(
     val id: Int,
-    val score: Int,
+    val score: Short,
     val comment: String?,
     val createdAt: Instant,
     val userDisplayName: String?,
